TheBackEnd/requirements.txt
absl-py==0.11.0
appdirs==1.4.3
asgiref==3.3.1
astor==0.8.1
CacheControl==0.12.6
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.5
django-numpy-json-encoder==0.1.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
html5lib==1.0.1
idna==2.8
importlib-metadata==3.4.0
ipaddr==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lockfile==0.12.2
Markdown==3.3.3
msgpack==0.6.2
numpy==1.19.5
opt-einsum==3.3.0
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
progress==1.5
protobuf==3.14.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
wrapt==1.12.1
zipp==3.4.0
